Update map review checklist and add another checkpoint placement graphic
Just treating this as both a checklist and a bit of an infodump for now
to get this information written down for people to reference right away.
Later it would probably be good to make the checklist item descriptions
very minimal and then have tooltips with extra info as a refresher.